Teach GitRepository.GetLatestTagHash() relativity
GetLatestTagHash() uses rev-list with all tags, and as such ends up
returning the commit with the latest date across the entire repo which
also has a tag. This differs from GetLatestTag(), which gets the latest
tag that is either on the current commit or any predecessor.

Teach GetLatestTagHash to only look at the current commit or any
predecessor to match the behaviour of GetLatestTag. This also fixes
IsTagCommit() and BuildContext.IsTagCommit, as they use
GetLatestTagHash() to determine if the current commit is tagged.